# Add AVX implementation for Global Average Pooling layer.