--- a/benchmarks/benchmarks.cpp
+++ b/benchmarks/benchmarks.cpp
@@ -1,0 +1,10 @@
+#ifndef _CRT_SECURE_NO_WARNINGS
+#define _CRT_SECURE_NO_WARNINGS
+#endif
+#include "benchmark/benchmark.h"
+#include "tiny_dnn/tiny_dnn.h"
+
+#include "bm_global_avepool.h"
+using namespace tiny_dnn::benchmarks;
+
+BENCHMARK_MAIN();