--- a/tiny_dnn/core/kernels/global_avepool_grad_op.h
+++ b/tiny_dnn/core/kernels/global_avepool_grad_op.h
@@ -30,19 +30,12 @@ class GlobalAvePoolGradOp : public core::OpKernel {
 
     const core::backend_t engine = context.engine();
 
-    if (engine == core::backend_t::internal) {
-      kernels::global_avepool_grad_op_internal(prev_delta, curr_delta, params,
-                                               context.parallelize());
-    } else if (engine == core::backend_t::avx) {
+    if (engine == core::backend_t::avx) {
       kernels::global_avepool_grad_op_avx(prev_delta, curr_delta, params,
                                           context.parallelize());
-    } else if (engine == core::backend_t::nnpack) {
-      // fallback to internal implementation as nnpack implementation
-      // is not available
+    } else {
       kernels::global_avepool_grad_op_internal(prev_delta, curr_delta, params,
                                                context.parallelize());
-    } else {
-      throw nn_error("Not supported engine: " + to_string(engine));
     }
   }
 };