… CMakeLists.txt
… benchmarks/CMakeLists.txt
… benchmarks/bm_global_avepool.h
… cmake/summary.cmake
… test/test_global_average_pooling_layer.h
… tiny_dnn/util/macro.h
… tiny_dnn/util/product.h