@karandesai-96 2017-05-06
@beru I think we should decide specific naming conventions for these vector blocks. Just for the sake of consistency. Right now to avoid confusion, I preferred suffixing such variables with a `_m`.