@beru 2017-05-06
@karandesai-96 I say NAY to Systems Hungarian notation.