@karandesai-96 2017-05-10
@beru I have worked with the review comments. Please have a look once more.
@edgarriba some patterns are different than other layers - such as fallback, macro protection etc. if this works well, we can make it uniform throughout.