@beru 2017-05-11
@karandesai-96 
> @beru I have worked with the review comments. Please have a look once more.

I checked the code but you still haven't written it in proper ways.
Though I'm unsure how further we should optimize this implementation.

Did you write sample code that uses  global average pooling layer?
How do you measure the performance of this AVX implementation for Global Average Pooling layer?