@karandesai-96 2017-05-13
@beru: Thanks for the extensive reviews. It helped me thinking about these kinds of optimizations beyond the horizon. It will certainly help me for next turns. I will proceed with loop unrolling now. 

> Did you write sample code that uses global average pooling layer?
> How do you measure the performance of this AVX implementation for Global Average Pooling layer?

I wrote a small benchmark myself, based on the code snippet I wrote in my [GSoC Proposal (Page 6)](https://docs.google.com/document/d/18RtcDatYzZ1gpztFUM8K2MUaYfIAwvVFIJKCR-xf6BE/edit).

I will share the execution timings for internal backend, AVX without unrolling and AVX with unrolling soon.