@beru 2017-05-15
@karandesai-96 
> I wrote a small benchmark myself, based on the code snippet I wrote in my GSoC Proposal (Page 6). I will share the execution timings for internal backend, AVX without unrolling and AVX with unrolling soon.

If you include the benchmark program to this PR, other people can also run it and check the performance.