@karandesai-96 2017-05-18
@beru Performed loop unrolling and added simple benchmarking code.
Here are the results:
![image](https://cloud.githubusercontent.com/assets/10494087/26211587/696a2072-3c10-11e7-916f-a9b767fc7d8f.png)

Also I cleaned up and rearranged my commit history to hope that this PR won't get squashed. xD