@karandesai-96 2017-05-20
@beru working on your suggestions to use `VECTORIZE_TYPE::add` slows down a bit due to calling overhead. (accumulate -> add -> _mm256_add)
```plaintext
// before:
------------------------------------------------------------------------------------------
Benchmark                                                   Time           CPU Iterations
------------------------------------------------------------------------------------------
bm_global_avepool_forward_internal/iterations:1000    5712979 ns      41563 ns       1000
bm_global_avepool_forward_avx/iterations:1000          716491 ns      44371 ns       1000

// after:
------------------------------------------------------------------------------------------
Benchmark                                                   Time           CPU Iterations
------------------------------------------------------------------------------------------
bm_global_avepool_forward_internal/iterations:1000    5190217 ns      52399 ns       1000
bm_global_avepool_forward_avx/iterations:1000         1056757 ns      45231 ns       1000

```