@karandesai-96 2017-05-20
On the other hand, removing excess if conditions from for loops and breaking the for loop into three sequential for loops speeds up a bit:
```plaintext
------------------------------------------------------------------------------------------
Benchmark                                                   Time           CPU Iterations
------------------------------------------------------------------------------------------
bm_global_avepool_forward_internal/iterations:1000    5071432 ns      46034 ns       1000
bm_global_avepool_forward_avx/iterations:1000          646088 ns      46132 ns       1000
```