@karandesai-96 2017-05-20
Instead of template metaprogramming, I suggest using macros. I tried this out:

```c++
#if defined(CNN_USE_AVX)
#ifdef CNN_USE_DOUBLE
#define __M256 __m256d
#define _MM256_ADD _mm256_add_pd

#else

#define __M256 __m256
#define _MM256_ADD _mm256_add_ps
#endif
```

This won't create an extra function call overhead like that suggested through template metaprogramming. And macros help to reduce the code as well as we achieve a speedup as compared to earlier case of 780k - 800k nanoseconds timings.

Results:
```plaintext
------------------------------------------------------------------------------------------
Benchmark                                                   Time           CPU Iterations
------------------------------------------------------------------------------------------
bm_global_avepool_forward_internal/iterations:1000    5136436 ns      48891 ns       1000
bm_global_avepool_forward_avx/iterations:1000          644082 ns      46916 ns       1000
```