@Randl 2017-05-20
I'm not sure that would work project-wide, but it might be convenient, unless we are forced to define some 1000 LoC of macros :)