@karandesai-96 2017-05-20
@Randl I think we shall require only certain commonly used ones, besides only one set of them would exist at a time due to `CNN_USE_AVX` / `CNN_USE_DOUBLE` etc. protections.