@beru 2017-05-20
> Instead of template metaprogramming, I suggest using macros.

I don't think preprocessor macros should be used often (though I use it when I have to)
There is `__forceinline` keyword for VC++, and `__attribute__((always_inline))` for GCC.
So if using these will remove extra function calls, we should stick with TMP.

```cpp
#ifdef _MSC_VER
#define MUSTINLINE __forceinline
#else
#define MUSTINLINE __attribute__((always_inline)) inline
#endif
```

I used `VECTORIZE_TYPE` macro in `tiny_dnn/util/product.h` since my C++ TMP skills are not good.

> I think we shall require only certain commonly used ones, besides only one set of them would exist at a time due to CNN_USE_AVX / CNN_USE_DOUBLE etc. protections.

I'd like this library to support multiple numeric data types in future. But using macros in that way limits the possibilties.