@beru 2017-05-20
@karandesai-96 

For `tiny_dnn_benchmarks` VC++ project, we need this include directory setting. `..\benchmark-src\include\`

By adding it, I could successfully build the project.